Create DatabaseException for integrityDatabase exception and adjusts

+++ ws_springBoot3/src/main/java/br/com/ctfera/course/services/UserService.java
@@ -2,10 +2,12 @@ package br.com.ctfera.course.services;
 
 import br.com.ctfera.course.entities.User;
 import br.com.ctfera.course.repositories.UserRepository;
+import br.com.ctfera.course.services.exceptions.DatabaseException;
 import br.com.ctfera.course.services.exceptions.ResourceNotFoundException;
 import ch.qos.logback.core.LogbackException;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
@@ -35,11 +37,13 @@ public class UserService {
    }
 
    public void delete(Long id){
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
         try{
             userRepository.deleteById(id);
-        }catch(EmptyResultDataAccessException e){
-           e.printStackTrace();
-
+        }catch(DataIntegrityViolationException e){
+           throw new DatabaseException(e.getMessage());
         }
    }
